refactor(event): split service.go below 1000 lines
Move alarm, attendee, and other helpers into sibling files
in package event. No behavior change.

Related: #652
