+++ internal/event/service.go
@@ -3,9 +3,7 @@ package event
 import (
 	"context"
 	"database/sql"
-	"errors"
 	"fmt"
-	"log"
 	"strings"
 	"time"
 
@@ -13,27 +11,10 @@ import (
 
 	"github.com/douglasdemoura/chroncal/internal/calendaraccess"
 	"github.com/douglasdemoura/chroncal/internal/duration"
-	"github.com/douglasdemoura/chroncal/internal/hydrate"
 	"github.com/douglasdemoura/chroncal/internal/model"
 	"github.com/douglasdemoura/chroncal/internal/storage"
 	"github.com/douglasdemoura/chroncal/internal/timeutil"
 )
-
-type SearchParams struct {
-	Query      string
-	CalendarID int64  // 0 = all calendars
-	From       string // RFC3339 or empty
-	To         string // RFC3339 or empty
-	Status     string // empty = all
-}
-
-type ExportParams struct {
-	CalendarID int64  // 0 = all
-	From       string // RFC3339 or empty
-	To         string // RFC3339 or empty
-	Category   string // empty = all
-	Status     string // empty = all
-}
 
 type Service struct {
 	db *sql.DB
@@ -212,110 +193,6 @@ func (p *UpsertParams) applyDefaults() {
 func (p *UpdateParams) applyDefaults() {
 	applyEventDefaults(&p.Status, &p.Transp, &p.Class)
 	normalizeEventTimes(&p.StartTime, &p.EndTime, p.AllDay)
-}
-
-func (s *Service) CountByCalendar(ctx context.Context, calendarID int64) (int64, error) {
-	return s.q.CountEventsByCalendar(ctx, calendarID)
-}
-
-func (s *Service) ListByDateRange(ctx context.Context, from, to time.Time) ([]Event, error) {
-	rows, err := s.q.ListEventsByDateRange(ctx, storage.ListEventsByDateRangeParams{
-		StartTime: to.UTC().Format(time.RFC3339),   // start_time < to
-		EndTime:   from.UTC().Format(time.RFC3339), // end_time > from
-	})
-	if err != nil {
-		return nil, err
-	}
-	events := fromStorageSlice(rows)
-	s.populateCategories(ctx, events)
-	return events, nil
-}
-
-func (s *Service) ListByCalendarAndDateRange(ctx context.Context, calID int64, from, to time.Time) ([]Event, error) {
-	rows, err := s.q.ListEventsByCalendarAndDateRange(ctx, storage.ListEventsByCalendarAndDateRangeParams{
-		CalendarID: calID,
-		StartTime:  to.UTC().Format(time.RFC3339),   // start_time < to
-		EndTime:    from.UTC().Format(time.RFC3339), // end_time > from
-	})
-	if err != nil {
-		return nil, err
-	}
-	events := fromStorageSlice(rows)
-	s.populateCategories(ctx, events)
-	return events, nil
-}
-
-func (s *Service) Search(ctx context.Context, p SearchParams) ([]Event, error) {
-	ftsQuery := storage.FTSQuery(p.Query)
-	if ftsQuery == "" {
-		return []Event{}, nil
-	}
-	rows, err := s.q.SearchEventsFTS(ctx, ftsQuery, p.CalendarID, p.From, p.To, p.Status)
-	if err != nil {
-		return nil, fmt.Errorf("search events: %w", err)
-	}
-	events := fromStorageSlice(rows)
-	s.populateCategories(ctx, events)
-	return events, nil
-}
-
-func (s *Service) ExportFiltered(ctx context.Context, p ExportParams) ([]Event, error) {
-	rows, err := s.q.ListEventsForExport(ctx, storage.EventFilterParams{
-		CalendarID:   p.CalendarID,
-		FromTime:     p.From,
-		ToTime:       p.To,
-		Category:     p.Category,
-		FilterStatus: p.Status,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("export events: %w", err)
-	}
-	events := fromStorageSlice(rows)
-	s.populateCategories(ctx, events)
-	return events, nil
-}
-
-func (s *Service) ListOverridesByUID(ctx context.Context, uid string) ([]Event, error) {
-	rows, err := s.q.ListOverridesByUID(ctx, uid)
-	if err != nil {
-		return nil, err
-	}
-	events := fromStorageSlice(rows)
-	s.populateCategories(ctx, events)
-	return events, nil
-}
-
-func (s *Service) Get(ctx context.Context, id int64) (Event, error) {
-	r, err := s.q.GetEvent(ctx, id)
-	if err != nil {
-		return Event{}, err
-	}
-	e := FromStorage(r)
-	s.populateSingleCategories(ctx, &e)
-	return e, nil
-}
-
-func (s *Service) GetByUID(ctx context.Context, uid string) (Event, error) {
-	r, err := s.q.GetEventByUID(ctx, uid)
-	if err != nil {
-		return Event{}, err
-	}
-	e := FromStorage(r)
-	s.populateSingleCategories(ctx, &e)
-	return e, nil
-}
-
-func (s *Service) GetByUIDAndRecurrenceID(ctx context.Context, uid, recurrenceID string) (Event, error) {
-	r, err := s.q.GetEventByUIDAndRecurrenceID(ctx, storage.GetEventByUIDAndRecurrenceIDParams{
-		Uid:          uid,
-		RecurrenceID: recurrenceID,
-	})
-	if err != nil {
-		return Event{}, err
-	}
-	e := FromStorage(r)
-	s.populateSingleCategories(ctx, &e)
-	return e, nil
 }
 
 // markDirtyByID looks up an event by ID and marks its sync resource as dirty.
@@ -692,344 +569,6 @@ func (s *Service) UpsertByUID(ctx context.Context, p UpsertParams) (Event, error
 	}
 	e.Categories = p.Categories
 	return e, nil
-}
-
-// ErrHasOverrides is returned when a delete targets a recurring master
-// event that has override instances. Use DeleteSeries instead.
-var ErrHasOverrides = fmt.Errorf("event has overrides: use DeleteSeries to delete the entire series")
-
-func (s *Service) Delete(ctx context.Context, id int64) error {
-	r, err := s.q.GetEvent(ctx, id)
-	if err != nil {
-		return err
-	}
-	evt := FromStorage(r)
-
-	if err := calendaraccess.EnsureWritable(ctx, s.q, evt.CalendarID, "VEVENT"); err != nil {
-		return err
-	}
-
-	// If this is a recurring master, check for overrides. RDATE-only masters
-	// (no RRULE) are recurring too, so guard on either rule or RDATEs (#415).
-	if (evt.RecurrenceRule != "" || evt.RDates != "") && evt.RecurrenceID == "" {
-		overrides, err := s.q.ListOverridesByUID(ctx, evt.UID)
-		if err != nil {
-			return fmt.Errorf("check overrides: %w", err)
-		}
-		if len(overrides) > 0 {
-			return ErrHasOverrides
-		}
-	}
-
-	// If this is a standalone event (no recurrence or a solo master), create
-	// a tombstone so the sync engine can send a DELETE to the server. The
-	// tombstone and the soft-delete commit together: if the tombstone write
-	// fails the soft-delete rolls back, so the next sync can never DELETE a
-	// still-live event from the server (issue #107).
-	if evt.RecurrenceID == "" {
-		tx, err := s.db.BeginTx(ctx, nil)
-		if err != nil {
-			return fmt.Errorf("begin tx: %w", err)
-		}
-		defer tx.Rollback()
-		qtx := s.q.WithTx(tx)
-		if _, err := storage.CreateTombstoneIfSynced(ctx, tx, evt.CalendarID, evt.UID); err != nil {
-			return fmt.Errorf("create tombstone: %w", err)
-		}
-		if err := qtx.SoftDeleteEvent(ctx, id); err != nil {
-			return fmt.Errorf("soft-delete event: %w", err)
-		}
-		return tx.Commit()
-	}
-
-	// If this is an override, add EXDATE to the master so the instance
-	// doesn't reappear on next expansion. The master's sync resource
-	// becomes dirty (modified EXDATE), not the override.
-	if evt.RecurrenceID != "" {
-		tx, err := s.db.BeginTx(ctx, nil)
-		if err != nil {
-			return fmt.Errorf("begin tx: %w", err)
-		}
-		defer tx.Rollback()
-		qtx := s.q.WithTx(tx)
-
-		master, err := qtx.GetEventByUID(ctx, evt.UID)
-		// A genuine lookup error (e.g. SQLITE_BUSY) must not collapse into the
-		// "no master" path: that would soft-delete the override while skipping
-		// its EXDATE/provenance bookkeeping, resurrecting the occurrence via
-		// series expansion (issue #412). Only a missing master (ErrNoRows)
-		// legitimately skips the bookkeeping.
-		if err != nil && !errors.Is(err, sql.ErrNoRows) {
-			return fmt.Errorf("get master: %w", err)
-		}
-		if err == nil {
-			existing := ParseTimeList(storage.NullableToString(master.Exdates))
-			recIDTime, parseErr := timeutil.ParseRecurrenceID(evt.RecurrenceID)
-			if parseErr != nil {
-				// A malformed recurrence_id can't be excluded from the
-				// master, so soft-deleting the override would resurrect the
-				// occurrence via series expansion. Fail loudly instead — the
-				// restore path treats the same parse failure as fatal.
-				return fmt.Errorf("parse recurrence_id %q: %w", evt.RecurrenceID, parseErr)
-			}
-			// All-day masters store recurrence_ids as full RFC 3339, so
-			// ParseRecurrenceID yields a UTC-located time. Re-tag it as
-			// date-only so the EXDATE serializes as VALUE=DATE matching
-			// DTSTART;VALUE=DATE on export (RFC 5545 §3.8.5.1, issue #221).
-			if master.AllDay == 1 {
-				recIDTime = timeutil.AsDateOnly(recIDTime)
-			}
-			existing = append(existing, recIDTime)
-			if err := qtx.UpdateEventExdates(ctx, storage.UpdateEventExdatesParams{
-				Exdates: storage.StringToNullable(SerializeTimeList(existing)),
-				ID:      master.ID,
-			}); err != nil {
-				return fmt.Errorf("update exdates: %w", err)
-			}
-			// Record provenance so restore knows this EXDATE was
-			// delete-added (and may be stripped) rather than imported.
-			if err := qtx.RecordEventExdateDelete(ctx, storage.RecordEventExdateDeleteParams{
-				CalendarID:   master.CalendarID,
-				Uid:          evt.UID,
-				RecurrenceID: evt.RecurrenceID,
-			}); err != nil {
-				return fmt.Errorf("record exdate delete: %w", err)
-			}
-		}
-
-		if err := qtx.SoftDeleteEvent(ctx, id); err != nil {
-			return fmt.Errorf("soft-delete event: %w", err)
-		}
-		// Mark the master dirty — its EXDATE was modified — inside the same
-		// transaction so a failed mark rolls the EXDATE change back rather than
-		// committing a change that is never pushed (issue #107).
-		if err := storage.MarkResourceDirty(ctx, tx, evt.CalendarID, evt.UID, "event"); err != nil {
-			return fmt.Errorf("mark resource dirty: %w", err)
-		}
-		return tx.Commit()
-	}
-
-	// Unreachable: RecurrenceID is either "" (handled above) or non-empty.
-	return s.q.SoftDeleteEvent(ctx, id)
-}
-
-// DeleteInstance excludes a single occurrence of a recurring event by adding
-// an EXDATE to the master. instanceTime is the StartTime of the occurrence.
-func (s *Service) DeleteInstance(ctx context.Context, uid string, instanceTime time.Time) error {
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin tx: %w", err)
-	}
-	defer tx.Rollback()
-	qtx := s.q.WithTx(tx)
-
-	// Read the master inside the transaction so the EXDATE list we recompute
-	// reflects a concurrent writer's changes (issue #116). Reading outside the
-	// tx let a second instance-delete clobber the first one's EXDATE.
-	master, err := qtx.GetEventByUID(ctx, uid)
-	if err != nil {
-		return fmt.Errorf("get master: %w", err)
-	}
-	if err := calendaraccess.EnsureWritable(ctx, qtx, master.CalendarID, "VEVENT"); err != nil {
-		return err
-	}
-
-	existing := ParseTimeList(storage.NullableToString(master.Exdates))
-	exdate := instanceTime.UTC()
-	// For an all-day master, tag the EXDATE as date-only so it serializes as
-	// VALUE=DATE matching DTSTART;VALUE=DATE on export; otherwise a strict
-	// CalDAV server ignores the mismatched DATE-TIME EXDATE and the deleted
-	// occurrence reappears (RFC 5545 §3.8.5.1, issue #221).
-	if master.AllDay == 1 {
-		exdate = timeutil.AsDateOnly(exdate)
-	}
-	existing = append(existing, exdate)
-	if err := qtx.UpdateEventExdates(ctx, storage.UpdateEventExdatesParams{
-		Exdates: storage.StringToNullable(SerializeTimeList(existing)),
-		ID:      master.ID,
-	}); err != nil {
-		return fmt.Errorf("update exdates: %w", err)
-	}
-
-	recID := instanceTime.UTC().Format(time.RFC3339)
-	override, oErr := qtx.GetEventByUIDAndRecurrenceID(ctx, storage.GetEventByUIDAndRecurrenceIDParams{
-		Uid:          uid,
-		RecurrenceID: recID,
-	})
-	if oErr == nil {
-		if err := qtx.SoftDeleteEvent(ctx, override.ID); err != nil {
-			return fmt.Errorf("soft-delete override: %w", err)
-		}
-	}
-
-	// Log the EXDATE-based delete so the trash view can surface it.
-	// ON CONFLICT upserts deleted_at, so deleting the same instance twice
-	// keeps exactly one log row with the latest timestamp.
-	if err := qtx.RecordEventExdateDelete(ctx, storage.RecordEventExdateDeleteParams{
-		CalendarID:   master.CalendarID,
-		Uid:          uid,
-		RecurrenceID: recID,
-	}); err != nil {
-		return fmt.Errorf("record exdate delete: %w", err)
-	}
-
-	if err := tx.Commit(); err != nil {
-		return err
-	}
-	_ = storage.MarkResourceDirty(ctx, s.db, master.CalendarID, uid, "event")
-	return nil
-}
-
-// DeleteFromInstance truncates a recurring series so that instances at or
-// after instanceTime are removed. It sets UNTIL on the RRULE. It soft-deletes
-// any overrides at or after the cutoff. It records the pre-truncation
-// RRULE in event_truncate_deletes so the trash view can restore it atomically.
-func (s *Service) DeleteFromInstance(ctx context.Context, uid string, instanceTime time.Time) error {
-	_, err := s.deleteFromInstance(ctx, uid, instanceTime)
-	return err
-}
-
-// softDeleteOverridesAndRecordTruncation trims the master's post-cutoff RDATEs.
-// It hides every live override at or after cutoff. It records the truncation
-// so the trash view can restore it. It captures the recurrence_ids it hides
-// and the RDATEs it drops BEFORE it removes them. Restore then re-shows
-// exactly those overrides and re-adds exactly those RDATEs.
-//
-// It does not restore an override the user deleted on its own (issue #287).
-// It does not leave a post-cutoff RDATE to reappear on the next expansion
-// (issue #463, since rrule-go expands RDATEs independently of the RRULE UNTIL
-// bound). Pairs with restoreTruncatedOverrides / restoreTruncatedRDates.
-//
-// prevRRule is the master's pre-truncation RRULE. The caller passes it because
-// it overwrote the master's recurrence_rule in the DB before this runs.
-func softDeleteOverridesAndRecordTruncation(ctx context.Context, qtx *storage.Queries, master storage.Event, instanceTime time.Time, prevRRule string) error {
-	uid := master.Uid
-	cutoff := instanceTime.UTC().Format(time.RFC3339)
-
-	removedRDates, err := trimMasterRDatesAtOrAfter(ctx, qtx, master, instanceTime)
-	if err != nil {
-		return err
-	}
-
-	hidden, err := qtx.ListLiveOverrideRecurrenceIDsAtOrAfter(ctx, storage.ListLiveOverrideRecurrenceIDsAtOrAfterParams{
-		Uid:          uid,
-		RecurrenceID: cutoff,
-	})
-	if err != nil {
-		return fmt.Errorf("list future overrides: %w", err)
-	}
-	hiddenList := strings.Join(hidden, ",")
-
-	if err := qtx.SoftDeleteOverridesAtOrAfter(ctx, storage.SoftDeleteOverridesAtOrAfterParams{
-		Uid:          uid,
-		RecurrenceID: cutoff,
-	}); err != nil {
-		return fmt.Errorf("soft-delete future overrides: %w", err)
-	}
-
-	if err := qtx.RecordEventTruncateDelete(ctx, storage.RecordEventTruncateDeleteParams{
-		CalendarID:      master.CalendarID,
-		Uid:             uid,
-		CutoffTime:      cutoff,
-		PreviousRrule:   prevRRule,
-		HiddenOverrides: &hiddenList,
-		RemovedRdates:   &removedRDates,
-	}); err != nil {
-		return fmt.Errorf("record truncate: %w", err)
-	}
-	return nil
-}
-
-// trimMasterRDatesAtOrAfter rewrites the master's rdates column to keep only the
-// occurrences strictly before instanceTime, and returns the dropped ones
-// serialized for the truncation log. It only issues an UPDATE when something
-// changes, so a master with no post-cutoff RDATEs is left untouched.
-func trimMasterRDatesAtOrAfter(ctx context.Context, qtx *storage.Queries, master storage.Event, instanceTime time.Time) (string, error) {
-	rdates := ParseTimeList(storage.NullableToString(master.Rdates))
-	if len(rdates) == 0 {
-		return "", nil
-	}
-	kept := make([]time.Time, 0, len(rdates))
-	var removed []time.Time
-	for _, rd := range rdates {
-		if rd.Before(instanceTime) {
-			kept = append(kept, rd)
-		} else {
-			removed = append(removed, rd)
-		}
-	}
-	if len(removed) == 0 {
-		return "", nil
-	}
-	if err := qtx.UpdateEventRdates(ctx, storage.UpdateEventRdatesParams{
-		Rdates: storage.StringToNullable(SerializeTimeList(kept)),
-		ID:     master.ID,
-	}); err != nil {
-		return "", fmt.Errorf("trim rdates: %w", err)
-	}
-	return SerializeTimeList(removed), nil
-}
-
-// deleteFromInstance performs the truncation and returns the master's
-// updated_at as written by this operation, read back inside the same
-// transaction. The in-tx value (rather than a post-commit read) closes a
-// TOCTOU window. A concurrent writer that edits the master after our commit
-// but before a separate read would otherwise have its updated_at captured as
-// the undo baseline. RestoreUndo would then clobber that edit.
-func (s *Service) deleteFromInstance(ctx context.Context, uid string, instanceTime time.Time) (string, error) {
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return "", fmt.Errorf("begin tx: %w", err)
-	}
-	defer tx.Rollback()
-	qtx := s.q.WithTx(tx)
-
-	// Read the master inside the transaction so the RRULE we truncate reflects
-	// a concurrent writer's edits rather than a pre-transaction snapshot
-	// (issue #116).
-	master, err := qtx.GetEventByUID(ctx, uid)
-	if err != nil {
-		return "", fmt.Errorf("get master: %w", err)
-	}
-	if err := calendaraccess.EnsureWritable(ctx, qtx, master.CalendarID, "VEVENT"); err != nil {
-		return "", err
-	}
-
-	// An RDATE-only master (no RRULE) has no recurrence rule to truncate.
-	// Synthesizing an "UNTIL=..." here would be an invalid bare RRULE that fails
-	// to parse, collapsing the whole series to its DTSTART (issue #414). Leave
-	// the rule NULL; the future-override soft-delete and truncation record below
-	// still apply.
-	prevRRule := storage.NullableToString(master.RecurrenceRule)
-	if prevRRule != "" {
-		until := instanceTime.UTC().Add(-time.Second)
-		rule := setRRuleUntil(prevRRule, until, master.AllDay == 1)
-		if err := qtx.UpdateEventRecurrenceRule(ctx, storage.UpdateEventRecurrenceRuleParams{
-			RecurrenceRule: storage.StringToNullable(rule),
-			ID:             master.ID,
-		}); err != nil {
-			return "", fmt.Errorf("update rrule: %w", err)
-		}
-	}
-
-	if err := softDeleteOverridesAndRecordTruncation(ctx, qtx, master, instanceTime, prevRRule); err != nil {
-		return "", err
-	}
-
-	// Read the master's updated_at back inside the transaction so the value we
-	// return reflects exactly this truncation's write, with no chance of an
-	// interleaved external edit in between.
-	truncated, err := qtx.GetEventByUID(ctx, uid)
-	if err != nil {
-		return "", fmt.Errorf("read back master: %w", err)
-	}
-	postUpdated := truncated.UpdatedAt
-
-	if err := tx.Commit(); err != nil {
-		return "", err
-	}
-	_ = storage.MarkResourceDirty(ctx, s.db, master.CalendarID, uid, "event")
-	return postUpdated, nil
 }
 
 // UpdateInstance creates or updates a per-occurrence override of a recurring
@@ -1445,1033 +984,4 @@ func setRRuleUntil(rule string, until time.Time, allDay bool) string {
 	}
 	out = append(out, untilStr)
 	return strings.Join(out, ";")
-}
-
-// DeleteSeries deletes a recurring master event and all its overrides.
-func (s *Service) DeleteSeries(ctx context.Context, uid string) error {
-	if err := s.ensureSeriesWritable(ctx, uid); err != nil {
-		return err
-	}
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin tx: %w", err)
-	}
-	defer tx.Rollback()
-	qtx := s.q.WithTx(tx)
-
-	// Look up the master to get calendarID for tombstone creation. The
-	// tombstone and the soft-delete commit together so a failed tombstone
-	// write can't leave a tombstone for a still-live series whose next sync
-	// would DELETE it from the server (issue #107).
-	if master, err := qtx.GetEventByUID(ctx, uid); err == nil {
-		if _, err := storage.CreateTombstoneIfSynced(ctx, tx, master.CalendarID, uid); err != nil {
-			return fmt.Errorf("create tombstone: %w", err)
-		}
-	}
-
-	if err := qtx.SoftDeleteEventsByUID(ctx, uid); err != nil {
-		return fmt.Errorf("soft-delete series: %w", err)
-	}
-	return tx.Commit()
-}
-
-// Alarm CRUD
-
-// buildAlarmsWithAttendees converts storage alarm rows into model.Alarm
-// values with attendees batch-loaded.
-func buildAlarmsWithAttendees(ctx context.Context, q *storage.Queries, rows []storage.EventAlarm) ([]model.Alarm, error) {
-	if len(rows) == 0 {
-		return nil, nil
-	}
-	alarmIDs := make([]int64, len(rows))
-	for i, r := range rows {
-		alarmIDs[i] = r.ID
-	}
-	// Load failures propagate: attendees feed content matching in
-	// ReplaceAlarms and X-properties feed export/sync pushes, so a silently
-	// degraded alarm set would corrupt merges or rewrite the server copy.
-	attRows, err := q.ListAlarmAttendeesByAlarmIDs(ctx, alarmIDs)
-	if err != nil {
-		return nil, fmt.Errorf("load alarm attendees: %w", err)
-	}
-	attMap := make(map[int64][]model.AlarmAttendee, len(rows))
-	for _, ar := range attRows {
-		attMap[ar.AlarmID] = append(attMap[ar.AlarmID], model.AlarmAttendee{
-			ID: ar.ID, Email: ar.Email, Name: storage.NullableToString(ar.Name),
-		})
-	}
-	alarms := make([]model.Alarm, len(rows))
-	for i, r := range rows {
-		alarms[i] = fromStorageAlarm(r)
-		alarms[i].Attendees = attMap[r.ID]
-	}
-	if err := storage.AttachAlarmXProperties(ctx, q, storage.OwnerTypeEventAlarm, alarms); err != nil {
-		return nil, err
-	}
-	return alarms, nil
-}
-
-func (s *Service) ListAlarms(ctx context.Context, eventID int64) ([]model.Alarm, error) {
-	rows, err := s.q.ListAlarmsByEventID(ctx, eventID)
-	if err != nil {
-		return nil, err
-	}
-	return buildAlarmsWithAttendees(ctx, s.q, rows)
-}
-
-// ListFireableAlarmsByEventIDs fetches the fireable alarms for multiple
-// event IDs in a single batch query. Returns a map of event ID to its list
-// of alarms. The query excludes preserved sync-only actions (issue #579):
-// the alarm check loop is the only caller, and a sync-only alarm must not
-// reach it.
-func (s *Service) ListFireableAlarmsByEventIDs(ctx context.Context, eventIDs []int64) (map[int64][]model.Alarm, error) {
-	if len(eventIDs) == 0 {
-		return nil, nil
-	}
-	alarmRows, err := s.q.ListFireableAlarmsByEventIDs(ctx, eventIDs)
-	if err != nil {
-		return nil, err
-	}
-	alarms, err := buildAlarmsWithAttendees(ctx, s.q, alarmRows)
-	if err != nil {
-		return nil, err
-	}
-	if len(alarms) == 0 {
-		return nil, nil
-	}
-	alarmMap := make(map[int64][]model.Alarm, len(eventIDs))
-	for _, a := range alarms {
-		alarmMap[a.EventID] = append(alarmMap[a.EventID], a)
-	}
-	return alarmMap, nil
-}
-
-// loadExistingAlarms loads alarms that already exist, with their attendees, for the given event.
-func loadExistingAlarms(ctx context.Context, qtx *storage.Queries, eventID int64) ([]model.Alarm, error) {
-	rows, err := qtx.ListAlarmsByEventID(ctx, eventID)
-	if err != nil {
-		return nil, fmt.Errorf("list existing alarms: %w", err)
-	}
-	return buildAlarmsWithAttendees(ctx, qtx, rows)
-}
-
-// matchAlarm tries to match a new alarm with stored ones by content.
-// Returns true and the index if matched, false otherwise. Rows whose
-// non-empty RFC 9074 UIDs differ are never paired. The UID identifies the
-// alarm. A content coincidence across different UIDs would attach
-// alarm_state to the wrong definition and churn UIDs on the server.
-func matchAlarm(existing []model.Alarm, matched []bool, a model.Alarm) (int, bool) {
-	for j, ex := range existing {
-		if matched[j] {
-			continue
-		}
-		if a.UID != "" && ex.UID != "" && a.UID != ex.UID {
-			continue
-		}
-		if a.ContentEqual(ex) {
-			return j, true
-		}
-	}
-	return 0, false
-}
-
-// alarmUID returns the UID this service stores for a. See
-// model.AlarmUIDForWrite for the rule, which the todo service shares.
-func alarmUID(a model.Alarm) string {
-	return model.AlarmUIDForWrite(a)
-}
-
-// matchAlarmByUID tries to match a new alarm with stored ones by
-// RFC 9074 UID. Use this as a fallback when content match fails. An edited
-// alarm (for example a changed trigger) then updates its row in place instead
-// of a delete and re-create. A delete would cascade away its alarm_state and
-// restore dismissed firings.
-func matchAlarmByUID(existing []model.Alarm, matched []bool, a model.Alarm) (int, bool) {
-	if a.UID == "" {
-		return 0, false
-	}
-	for j, ex := range existing {
-		if matched[j] || ex.UID == "" {
-			continue
-		}
-		if ex.UID == a.UID {
-			return j, true
-		}
-	}
-	return 0, false
-}
-
-// updateAlarmInPlace rewrites a UID-matched alarm's content on its stored
-// row. The row ID stays so alarm_state entries keyed to it survive.
-func updateAlarmInPlace(ctx context.Context, qtx *storage.Queries, eventID int64, a model.Alarm, ex model.Alarm) error {
-	// A rewrite to a sync-only action disables the alarm, but this code
-	// leaves alarm_state alone. The pending and snooze queries filter on
-	// the current action, so the state of the alarm comes back when a
-	// later pull restores a fireable action (issue #579).
-	ack, err := model.PrepareAlarmUpdate(a, ex)
-	if err != nil {
-		return err
-	}
-	if err := qtx.UpdateAlarmContentByID(ctx, storage.UpdateAlarmContentByIDParams{
-		Action:        a.Action,
-		TriggerValue:  a.TriggerValue,
-		Description:   storage.StringToNullable(a.Description),
-		Summary:       storage.StringToNullable(a.Summary),
-		Repeat:        int64(a.Repeat),
-		Duration:      storage.StringToNullable(a.Duration),
-		Related:       a.Related,
-		Acknowledged:  storage.StringToNullable(ack),
-		AttachUri:     storage.StringToNullable(a.AttachURI),
-		AttachFmttype: storage.StringToNullable(a.AttachFmtType),
-		AttachBinary:  a.AttachBinary,
-		ID:            ex.ID,
-		EventID:       eventID,
-	}); err != nil {
-		return fmt.Errorf("update alarm content: %w", err)
-	}
-	if err := qtx.DeleteAlarmAttendeesByAlarmID(ctx, ex.ID); err != nil {
-		return fmt.Errorf("delete alarm attendees: %w", err)
-	}
-	for _, att := range a.Attendees {
-		_, err := qtx.CreateAlarmAttendee(ctx, storage.CreateAlarmAttendeeParams{
-			AlarmID: ex.ID,
-			Email:   att.Email,
-			Name:    storage.StringToNullable(att.Name),
-		})
-		if err != nil {
-			return fmt.Errorf("create alarm attendee: %w", err)
-		}
-	}
-	if a.XProperties == nil || model.XPropsContentEqual(a.XProperties, ex.XProperties) {
-		return nil
-	}
-	return storage.ReplaceAlarmXProperties(ctx, qtx, storage.OwnerTypeEventAlarm, ex.ID, a.XProperties)
-}
-
-// syncMatchedAlarm syncs a matched alarm's UID and ACKNOWLEDGED state.
-func syncMatchedAlarm(ctx context.Context, qtx *storage.Queries, eventID int64, a model.Alarm, ex model.Alarm) error {
-	// Backfill the UID of a stored row that carries none. A preserved
-	// foreign alarm keeps its empty UID, so the backfill writes nothing
-	// for it (issue #586).
-	if uid := alarmUID(a); ex.UID == "" && uid != "" {
-		if err := qtx.UpdateAlarmUID(ctx, storage.UpdateAlarmUIDParams{
-			Uid: storage.StringToNullable(uid),
-			ID:  ex.ID,
-		}); err != nil {
-			return fmt.Errorf("backfill alarm uid: %w", err)
-		}
-	}
-	// Sync ACKNOWLEDGED if the incoming value differs (including clearing).
-	if a.Acknowledged != ex.Acknowledged && model.ValidateAcknowledged(a.Acknowledged) {
-		if err := qtx.UpdateAlarmAcknowledged(ctx, storage.UpdateAlarmAcknowledgedParams{
-			Acknowledged: storage.StringToNullable(a.Acknowledged),
-			ID:           ex.ID,
-			EventID:      eventID,
-		}); err != nil {
-			return fmt.Errorf("update alarm acknowledged: %w", err)
-		}
-	}
-	// X-properties are excluded from content matching; refresh them so a
-	// remote X-prop change still lands. nil means the caller has no X-prop
-	// knowledge (CLI flags, TUI fallback paths) — keep the stored rows; only
-	// a non-nil slice (import/sync always populates one) is authoritative.
-	if a.XProperties == nil || model.XPropsContentEqual(a.XProperties, ex.XProperties) {
-		return nil
-	}
-	return storage.ReplaceAlarmXProperties(ctx, qtx, storage.OwnerTypeEventAlarm, ex.ID, a.XProperties)
-}
-
-// isUniqueUIDViolation reports whether an insert failed on the global
-// alarm-UID unique index.
-func isUniqueUIDViolation(err error) bool {
-	if err == nil {
-		return false
-	}
-	msg := err.Error()
-	return strings.Contains(msg, "UNIQUE constraint failed") && strings.Contains(msg, ".uid")
-}
-
-// createNewAlarm creates a new alarm and its attendees. Alarm UIDs are
-// globally unique. Servers sometimes duplicate an event (same VALARM UIDs on
-// both copies), which would otherwise fail this event's sync forever. On
-// collision, mint a fresh local UID instead.
-func createNewAlarm(ctx context.Context, qtx *storage.Queries, eventID int64, a model.Alarm) error {
-	if err := model.CheckStorableAlarmAction(a.Action); err != nil {
-		return fmt.Errorf("create alarm: %w", err)
-	}
-	params := storage.CreateAlarmParams{
-		EventID:       eventID,
-		Uid:           storage.StringToNullable(alarmUID(a)),
-		Action:        a.Action,
-		TriggerValue:  a.TriggerValue,
-		Description:   storage.StringToNullable(a.Description),
-		Summary:       storage.StringToNullable(a.Summary),
-		Repeat:        int64(a.Repeat),
-		Duration:      storage.StringToNullable(a.Duration),
-		Related:       a.Related,
-		Acknowledged:  storage.StringToNullable(a.Acknowledged),
-		AttachUri:     storage.StringToNullable(a.AttachURI),
-		AttachFmttype: storage.StringToNullable(a.AttachFmtType),
-		AttachBinary:  a.AttachBinary,
-	}
-	row, err := qtx.CreateAlarm(ctx, params)
-	if isUniqueUIDViolation(err) {
-		// A server can send the same VALARM UID on two resources. Retry
-		// through the shared rule, so the retry does not stamp a minted
-		// UID on a preserved foreign alarm (issue #586). The rule mints
-		// for a fireable action and yields an empty value otherwise.
-		retry := a
-		retry.UID = ""
-		params.Uid = storage.StringToNullable(model.AlarmUIDForWrite(retry))
-		row, err = qtx.CreateAlarm(ctx, params)
-	}
-	if err != nil {
-		return fmt.Errorf("create alarm: %w", err)
-	}
-	for _, att := range a.Attendees {
-		_, err := qtx.CreateAlarmAttendee(ctx, storage.CreateAlarmAttendeeParams{
-			AlarmID: row.ID,
-			Email:   att.Email,
-			Name:    storage.StringToNullable(att.Name),
-		})
-		if err != nil {
-			return fmt.Errorf("create alarm attendee: %w", err)
-		}
-	}
-	if len(a.XProperties) == 0 {
-		return nil
-	}
-	return storage.ReplaceAlarmXProperties(ctx, qtx, storage.OwnerTypeEventAlarm, row.ID, a.XProperties)
-}
-
-// deleteUnmatchedAlarms deletes alarms that were not matched.
-func deleteUnmatchedAlarms(ctx context.Context, qtx *storage.Queries, existing []model.Alarm, matched []bool) error {
-	for j, ex := range existing {
-		if !matched[j] {
-			if err := qtx.DeleteAlarmByID(ctx, ex.ID); err != nil {
-				return fmt.Errorf("delete unmatched alarm: %w", err)
-			}
-		}
-	}
-	return nil
-}
-
-// ReplaceFireableAlarms replaces the fireable alarms of an event and
-// carries the stored sync-only rows forward. A caller that cannot state a
-// preserved action — the CLI --alarm flag — uses this method, so a
-// routine edit does not delete the VALARM of another client (issue #579).
-// A caller that must delete such a row calls ReplaceAlarms instead.
-// The read of the stored rows and the write share one transaction. A read
-// on its own connection could return a row a concurrent pull has already
-// deleted. The carry-over would then write that row again, and the next
-// push would restore the deleted VALARM on the server.
-func (s *Service) ReplaceFireableAlarms(ctx context.Context, eventID int64, alarms []model.Alarm) error {
-	if s.tx != nil {
-		return s.replaceFireableAlarms(ctx, eventID, alarms)
-	}
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin tx: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	if err := s.WithTx(tx).replaceFireableAlarms(ctx, eventID, alarms); err != nil {
-		return err
-	}
-	return tx.Commit()
-}
-
-// replaceFireableAlarms carries the stored sync-only rows forward. The
-// caller supplies the transaction.
-func (s *Service) replaceFireableAlarms(ctx context.Context, eventID int64, alarms []model.Alarm) error {
-	stored, err := s.ListAlarms(ctx, eventID)
-	if err != nil {
-		return fmt.Errorf("list stored alarms: %w", err)
-	}
-	return s.ReplaceAlarms(ctx, eventID, model.KeepSyncOnlyAlarms(stored, alarms))
-}
-
-// ClearSyncOnlyAlarms deletes every stored alarm the engine cannot fire and
-// keeps the rest. ReplaceFireableAlarms carries those rows forward, so a
-// --alarm edit alone can never remove one. This method gives the user a way
-// out on a local calendar (issue #593).
-//
-// The read of the stored rows and the write share one transaction, like
-// ReplaceFireableAlarms.
-func (s *Service) ClearSyncOnlyAlarms(ctx context.Context, eventID int64) error {
-	if s.tx != nil {
-		return s.clearSyncOnlyAlarms(ctx, eventID)
-	}
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin tx: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	if err := s.WithTx(tx).clearSyncOnlyAlarms(ctx, eventID); err != nil {
-		return err
-	}
-	return tx.Commit()
-}
-
-// clearSyncOnlyAlarms keeps the fireable stored rows. The caller supplies
-// the transaction.
-func (s *Service) clearSyncOnlyAlarms(ctx context.Context, eventID int64) error {
-	stored, err := s.ListAlarms(ctx, eventID)
-	if err != nil {
-		return fmt.Errorf("list stored alarms: %w", err)
-	}
-	return s.ReplaceAlarms(ctx, eventID, model.FireableAlarmsOnly(stored))
-}
-
-func (s *Service) ReplaceAlarms(ctx context.Context, eventID int64, alarms []model.Alarm) error {
-	if err := s.ensureEventWritable(ctx, eventID, 0); err != nil {
-		return err
-	}
-	return s.ReplaceAlarmsForSync(ctx, eventID, alarms)
-}
-
-// ReplaceAlarmsForSync applies an alarm set without the remote access and
-// component policy. It is reserved for the CalDAV sync engine, which
-// mirrors a server-originated VEVENT into the local cache whatever the
-// linked collection advertises. A user-originated edit must route through
-// ReplaceAlarms, so the policy holds.
-func (s *Service) ReplaceAlarmsForSync(ctx context.Context, eventID int64, alarms []model.Alarm) error {
-	// Prepare before the transaction opens. A standalone call then
-	// rejects a bad alarm without a write lock. A sync caller already
-	// holds its own transaction. See model.PrepareAlarmsForWrite.
-	alarms, err := model.PrepareAlarmsForWrite(alarms)
-	if err != nil {
-		return err
-	}
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-
-	if err := replaceAlarmsTx(ctx, qtx, eventID, alarms); err != nil {
-		return err
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, eventID)
-	return nil
-}
-
-// replaceAlarmsTx reconciles an event's alarms (content/UID match, in-place
-// edits, creates, deletes) using a tx-bound Queries. It opens no transaction so
-// callers can compose it with the event row write inside one transaction.
-func replaceAlarmsTx(ctx context.Context, qtx *storage.Queries, eventID int64, alarms []model.Alarm) error {
-	// Precondition: the caller prepares alarms with
-	// model.PrepareAlarmsForWrite. Both callers in this file do.
-	// Load existing alarms with attendees for content matching.
-	existing, err := loadExistingAlarms(ctx, qtx, eventID)
-	if err != nil {
-		return err
-	}
-
-	// Match incoming alarms against existing by content.
-	// Slice-based matching: each existing alarm can only match once (supports duplicates).
-	matched := make([]bool, len(existing))
-	var unmatched []model.Alarm
-	for _, a := range alarms {
-		if j, found := matchAlarm(existing, matched, a); found {
-			matched[j] = true
-			if err := syncMatchedAlarm(ctx, qtx, eventID, a, existing[j]); err != nil {
-				return err
-			}
-		} else {
-			unmatched = append(unmatched, a)
-		}
-	}
-
-	// Second pass: alarms whose content changed but whose RFC 9074 UID is
-	// stable are the same alarm edited, not a new one. Update in place so
-	// the row ID — and the alarm_state rows hanging off it — survive.
-	for _, a := range unmatched {
-		if j, found := matchAlarmByUID(existing, matched, a); found {
-			matched[j] = true
-			if err := updateAlarmInPlace(ctx, qtx, eventID, a, existing[j]); err != nil {
-				return err
-			}
-		} else {
-			if err := createNewAlarm(ctx, qtx, eventID, a); err != nil {
-				return err
-			}
-		}
-	}
-
-	// Delete existing alarms that were not matched (they were removed).
-	return deleteUnmatchedAlarms(ctx, qtx, existing, matched)
-}
-
-// replaceRelationsTx replaces an event's attendees and alarms using a tx-bound
-// Queries. The *WithRelations methods can then write both child collections
-// inside the same transaction as the event row.
-func replaceRelationsTx(ctx context.Context, qtx *storage.Queries, eventID int64, attendees []model.Attendee, alarms []model.Alarm) error {
-	// The *WithRelations methods prepare the attendees and the alarms at
-	// method entry.
-	if err := replaceAttendeesTx(ctx, qtx, eventID, attendees); err != nil {
-		return err
-	}
-	return replaceAlarmsTx(ctx, qtx, eventID, alarms)
-}
-
-// Attendee CRUD
-
-func (s *Service) ListAttendees(ctx context.Context, eventID int64) ([]model.Attendee, error) {
-	rows, err := s.q.ListAttendeesByEventID(ctx, eventID)
-	if err != nil {
-		return nil, err
-	}
-	attendees := make([]model.Attendee, len(rows))
-	for i, r := range rows {
-		attendees[i] = fromStorageAttendee(r)
-	}
-	return attendees, nil
-}
-
-func (s *Service) ReplaceAttendees(ctx context.Context, eventID int64, attendees []model.Attendee) error {
-	if err := s.ensureEventWritable(ctx, eventID, 0); err != nil {
-		return err
-	}
-	return s.ReplaceAttendeesForSync(ctx, eventID, attendees)
-}
-
-// ReplaceAttendeesForSync applies an attendee set without remote
-// access/component policy. It is reserved for the CalDAV sync engine, which
-// mirrors server-originated VEVENTs into the local cache regardless of the
-// linked collection's advertised write support. User-originated edits (for
-// example the TUI RSVP flow) must route through ReplaceAttendees so the
-// policy is enforced.
-func (s *Service) ReplaceAttendeesForSync(ctx context.Context, eventID int64, attendees []model.Attendee) error {
-	// Prepare before the transaction opens. A standalone call then
-	// rejects a bad attendee without a write lock. See
-	// model.PrepareAttendeesForWrite.
-	attendees, err := model.PrepareAttendeesForWrite(model.EventAttendee, attendees)
-	if err != nil {
-		return err
-	}
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-
-	if err := replaceAttendeesTx(ctx, qtx, eventID, attendees); err != nil {
-		return err
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, eventID)
-	return nil
-}
-
-// replaceAttendeesTx replaces an event's attendees using a tx-bound Queries. It
-// opens no transaction so callers can compose it with the event row write
-// inside one transaction.
-func replaceAttendeesTx(ctx context.Context, qtx *storage.Queries, eventID int64, attendees []model.Attendee) error {
-	if err := qtx.DeleteAttendeesByEventID(ctx, eventID); err != nil {
-		return fmt.Errorf("delete attendees: %w", err)
-	}
-	for _, a := range attendees {
-		rsvp := ""
-		if a.RSVPRequested {
-			rsvp = "TRUE"
-		}
-		_, err := qtx.CreateAttendee(ctx, storage.CreateAttendeeParams{
-			EventID:       eventID,
-			Email:         a.Email,
-			Name:          storage.StringToNullable(a.Name),
-			RsvpStatus:    a.RSVPStatus,
-			Role:          a.Role,
-			Organizer:     storage.BoolToInt(a.Organizer),
-			Cutype:        storage.StringToNullable(a.CUType),
-			Rsvp:          storage.StringToNullable(rsvp),
-			SentBy:        storage.StringToNullable(a.SentBy),
-			DelegatedTo:   storage.StringToNullable(a.DelegatedTo),
-			DelegatedFrom: storage.StringToNullable(a.DelegatedFrom),
-			Member:        storage.StringToNullable(a.Member),
-			Dir:           storage.StringToNullable(a.Dir),
-			Language:      storage.StringToNullable(a.Language),
-		})
-		if err != nil {
-			return fmt.Errorf("create attendee: %w", err)
-		}
-	}
-	return nil
-}
-
-// Category CRUD
-
-func (s *Service) ListCategories(ctx context.Context, eventID int64) ([]string, error) {
-	rows, err := s.q.ListCategoriesByEventID(ctx, eventID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]string, len(rows))
-	for i, r := range rows {
-		out[i] = r.Category
-	}
-	return out, nil
-}
-
-func (s *Service) ListAllCategories(ctx context.Context) ([]string, error) {
-	return s.q.ListAllEventCategories(ctx)
-}
-
-func (s *Service) ReplaceCategories(ctx context.Context, eventID int64, categories []string) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-
-	if err := replaceCategoriesTx(ctx, qtx, eventID, categories); err != nil {
-		return err
-	}
-	if err := commit(); err != nil {
-		return fmt.Errorf("commit replace categories: %w", err)
-	}
-	return nil
-}
-
-// replaceCategoriesTx replaces an event's categories using a tx-bound Queries.
-// It does not open or commit a transaction, so callers can compose it with
-// other writes inside a single transaction.
-func replaceCategoriesTx(ctx context.Context, qtx *storage.Queries, eventID int64, categories []string) error {
-	if err := qtx.DeleteCategoriesByEventID(ctx, eventID); err != nil {
-		return fmt.Errorf("delete categories: %w", err)
-	}
-	for _, c := range categories {
-		_, err := qtx.CreateEventCategory(ctx, storage.CreateEventCategoryParams{
-			EventID:  eventID,
-			Category: c,
-		})
-		if err != nil {
-			return fmt.Errorf("create category: %w", err)
-		}
-	}
-	return nil
-}
-
-// Attachment CRUD
-
-func (s *Service) ListAttachments(ctx context.Context, eventID int64) ([]model.Attachment, error) {
-	rows, err := s.q.ListEventAttachmentsByEventID(ctx, eventID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]model.Attachment, len(rows))
-	for i, r := range rows {
-		out[i] = model.Attachment{ID: r.ID, URI: storage.NullableToString(r.Uri), FmtType: storage.NullableToString(r.Fmttype), Data: r.Data, Filename: storage.NullableToString(r.Filename)}
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceAttachments(ctx context.Context, eventID int64, attachments []model.Attachment) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteEventAttachmentsByEventID(ctx, eventID); err != nil {
-		return fmt.Errorf("delete attachments: %w", err)
-	}
-	for _, a := range attachments {
-		_, err := qtx.CreateEventAttachment(ctx, storage.CreateEventAttachmentParams{
-			EventID: eventID, Uri: storage.StringToNullable(a.URI), Fmttype: storage.StringToNullable(a.FmtType), Data: a.Data, Filename: storage.StringToNullable(a.Filename),
-		})
-		if err != nil {
-			return fmt.Errorf("create attachment: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, eventID)
-	return nil
-}
-
-// Comment CRUD
-
-func (s *Service) ListComments(ctx context.Context, eventID int64) ([]string, error) {
-	rows, err := s.q.ListEventCommentsByEventID(ctx, eventID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]string, len(rows))
-	for i, r := range rows {
-		out[i] = r.Text
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceComments(ctx context.Context, eventID int64, comments []string) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteEventCommentsByEventID(ctx, eventID); err != nil {
-		return fmt.Errorf("delete comments: %w", err)
-	}
-	for _, c := range comments {
-		_, err := qtx.CreateEventComment(ctx, storage.CreateEventCommentParams{
-			EventID: eventID, Text: c,
-		})
-		if err != nil {
-			return fmt.Errorf("create comment: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, eventID)
-	return nil
-}
-
-// Contact CRUD
-
-func (s *Service) ListContacts(ctx context.Context, eventID int64) ([]string, error) {
-	rows, err := s.q.ListEventContactsByEventID(ctx, eventID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]string, len(rows))
-	for i, r := range rows {
-		out[i] = r.Text
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceContacts(ctx context.Context, eventID int64, contacts []string) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteEventContactsByEventID(ctx, eventID); err != nil {
-		return fmt.Errorf("delete contacts: %w", err)
-	}
-	for _, c := range contacts {
-		_, err := qtx.CreateEventContact(ctx, storage.CreateEventContactParams{
-			EventID: eventID, Text: c,
-		})
-		if err != nil {
-			return fmt.Errorf("create contact: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, eventID)
-	return nil
-}
-
-// Resource CRUD
-
-func (s *Service) ListResources(ctx context.Context, eventID int64) ([]string, error) {
-	rows, err := s.q.ListEventResourcesByEventID(ctx, eventID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]string, len(rows))
-	for i, r := range rows {
-		out[i] = r.Text
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceResources(ctx context.Context, eventID int64, resources []string) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteEventResourcesByEventID(ctx, eventID); err != nil {
-		return fmt.Errorf("delete resources: %w", err)
-	}
-	for _, r := range resources {
-		_, err := qtx.CreateEventResource(ctx, storage.CreateEventResourceParams{
-			EventID: eventID, Text: r,
-		})
-		if err != nil {
-			return fmt.Errorf("create resource: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, eventID)
-	return nil
-}
-
-// Relation CRUD
-
-func (s *Service) ListRelations(ctx context.Context, eventID int64) ([]model.Relation, error) {
-	rows, err := s.q.ListEventRelationsByEventID(ctx, eventID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]model.Relation, len(rows))
-	for i, r := range rows {
-		out[i] = model.Relation{ID: r.ID, RelType: r.RelType, RelUID: r.RelUid}
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceRelations(ctx context.Context, eventID int64, relations []model.Relation) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteEventRelationsByEventID(ctx, eventID); err != nil {
-		return fmt.Errorf("delete relations: %w", err)
-	}
-	for _, r := range relations {
-		_, err := qtx.CreateEventRelation(ctx, storage.CreateEventRelationParams{
-			EventID: eventID, RelType: r.RelType, RelUid: r.RelUID,
-		})
-		if err != nil {
-			return fmt.Errorf("create relation: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, eventID)
-	return nil
-}
-
-// Converters
-
-// FromStorage maps a storage.Event row to a domain Event. It is exported so
-// recurrence expansion shares one mapper with the event service. The two then
-// cannot drift. Every new storage.Event column lands here once.
-func FromStorage(r storage.Event) Event {
-	var deletedAt *time.Time
-	if r.DeletedAt != nil && *r.DeletedAt != "" {
-		t := timeutil.ParseDateTime(*r.DeletedAt)
-		deletedAt = &t
-	}
-	return Event{
-		ID:             r.ID,
-		UID:            r.Uid,
-		CalendarID:     r.CalendarID,
-		Title:          r.Title,
-		Description:    storage.NullableToString(r.Description),
-		Location:       storage.NullableToString(r.Location),
-		StartTime:      timeutil.ParseDateTime(r.StartTime),
-		EndTime:        timeutil.ParseDateTime(r.EndTime),
-		AllDay:         r.AllDay == 1,
-		RecurrenceRule: storage.NullableToString(r.RecurrenceRule),
-		Timezone:       storage.NullableToString(r.Timezone),
-		Status:         r.Status,
-		Transp:         r.Transp,
-		Sequence:       r.Sequence,
-		Priority:       r.Priority,
-		Class:          r.Class,
-		URL:            storage.NullableToString(r.Url),
-		ConferenceURI:  r.ConferenceUri,
-		ExDates:        storage.NullableToString(r.Exdates),
-		RDates:         storage.NullableToString(r.Rdates),
-		RecurrenceID:   r.RecurrenceID,
-		Geo:            storage.NullableToString(r.Geo),
-		DurationValue:  storage.NullableToString(r.Duration),
-		DtStamp:        storage.NullableToString(r.Dtstamp),
-		CreatedAt:      timeutil.ParseDateTime(r.CreatedAt),
-		UpdatedAt:      timeutil.ParseDateTime(r.UpdatedAt),
-		DeletedAt:      deletedAt,
-	}
-}
-
-func fromStorageSlice(rows []storage.Event) []Event {
-	events := make([]Event, len(rows))
-	for i, r := range rows {
-		events[i] = FromStorage(r)
-	}
-	return events
-}
-
-func (s *Service) populateSingleCategories(ctx context.Context, e *Event) {
-	cats, err := s.ListCategories(ctx, e.ID)
-	if err != nil {
-		log.Printf("populateSingleCategories failed for event %d: %v", e.ID, err)
-		return
-	}
-	e.Categories = timeutil.JoinCategoryList(cats)
-}
-
-func (s *Service) populateCategories(ctx context.Context, events []Event) {
-	if len(events) == 0 {
-		return
-	}
-	ids := make([]int64, len(events))
-	for i := range events {
-		ids[i] = events[i].ID
-	}
-	rows, err := s.q.ListCategoriesByEventIDs(ctx, ids)
-	if err != nil {
-		log.Printf("populateCategories failed for %d events: %v", len(events), err)
-		return
-	}
-	catMap := make(map[int64][]string, len(events))
-	for _, r := range rows {
-		catMap[r.EventID] = append(catMap[r.EventID], r.Category)
-	}
-	for i := range events {
-		if cats, ok := catMap[events[i].ID]; ok {
-			events[i].Categories = timeutil.JoinCategoryList(cats)
-		}
-	}
-}
-
-// fromStorageAlarm converts an alarm row to the model value. It maps a
-// malformed stored action to model.UnsupportedAlarmAction, so every
-// reader holds a value the write rule accepts. See
-// model.NormalizeAlarmAction for the rule (issue #607).
-func fromStorageAlarm(r storage.EventAlarm) model.Alarm {
-	return model.Alarm{
-		ID:            r.ID,
-		EventID:       r.EventID,
-		UID:           storage.NullableToString(r.Uid),
-		Action:        model.NormalizeAlarmAction(r.Action),
-		TriggerValue:  r.TriggerValue,
-		Description:   storage.NullableToString(r.Description),
-		Summary:       storage.NullableToString(r.Summary),
-		Repeat:        int(r.Repeat),
-		Duration:      storage.NullableToString(r.Duration),
-		Related:       r.Related,
-		Acknowledged:  storage.NullableToString(r.Acknowledged),
-		AttachURI:     storage.NullableToString(r.AttachUri),
-		AttachBinary:  r.AttachBinary,
-		AttachFmtType: storage.NullableToString(r.AttachFmttype),
-	}
-}
-
-// X-Property CRUD
-
-func (s *Service) ListXProperties(ctx context.Context, eventID int64) ([]model.XProperty, error) {
-	rows, err := s.q.ListXPropertiesByOwner(ctx, storage.ListXPropertiesByOwnerParams{
-		OwnerType: "event", OwnerID: eventID,
-	})
-	if err != nil {
-		return nil, err
-	}
-	out := make([]model.XProperty, len(rows))
-	for i, r := range rows {
-		out[i] = model.XProperty{
-			ID: r.ID, OwnerType: r.OwnerType, OwnerID: r.OwnerID,
-			Name: r.Name, Value: r.Value, Params: r.Params,
-		}
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceXProperties(ctx context.Context, eventID int64, xprops []model.XProperty) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteXPropertiesByOwner(ctx, storage.DeleteXPropertiesByOwnerParams{
-		OwnerType: "event", OwnerID: eventID,
-	}); err != nil {
-		return fmt.Errorf("delete x-properties: %w", err)
-	}
-	for _, xp := range xprops {
-		if err := qtx.InsertXProperty(ctx, storage.InsertXPropertyParams{
-			OwnerType: "event", OwnerID: eventID,
-			Name: xp.Name, Value: xp.Value, Params: xp.Params,
-		}); err != nil {
-			return fmt.Errorf("insert x-property: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, eventID)
-	return nil
-}
-
-func fromStorageAttendee(r storage.EventAttendee) model.Attendee {
-	return model.Attendee{
-		ID:            r.ID,
-		EventID:       r.EventID,
-		Email:         r.Email,
-		Name:          storage.NullableToString(r.Name),
-		RSVPStatus:    r.RsvpStatus,
-		Role:          r.Role,
-		Organizer:     r.Organizer == 1,
-		CUType:        storage.NullableToString(r.Cutype),
-		RSVPRequested: strings.EqualFold(storage.NullableToString(r.Rsvp), "TRUE"),
-		SentBy:        storage.NullableToString(r.SentBy),
-		DelegatedTo:   storage.NullableToString(r.DelegatedTo),
-		DelegatedFrom: storage.NullableToString(r.DelegatedFrom),
-		Member:        storage.NullableToString(r.Member),
-		Dir:           storage.NullableToString(r.Dir),
-		Language:      storage.NullableToString(r.Language),
-	}
-}
-
-// Hydrate loads the transient relation slices — alarms, attendees, attachments,
-// comments, contacts, resources, relations, x-properties — onto e. They live in
-// side tables rather than the events row. A value straight out of Get/List
-// carries none of them. Any iCal built from it would omit them in silence.
-//
-// This is the single definition of "what a fully populated event is". Export
-// (CLI and file), CalDAV push, and the display paths all call it. A relation
-// added later then reaches every consumer at once. It fails on the first error
-// instead of a partial value. A caller that pushes an amputated record
-// overwrites the server copy with it.
-func (s *Service) Hydrate(ctx context.Context, e *Event) error {
-	return s.hydrate(ctx, e, true)
-}
-
-// HydrateBestEffort populates every relation it can and returns the joined
-// errors for the ones it could not. It does not stop at the first failure.
-//
-// This is for read-only display paths. One unreadable relation should degrade
-// that field alone. An early stop would leave every relation after it nil.
-// A caller that renders JSON cannot tell that apart from "there are none".
-// Any path that writes iCal must use Hydrate. A partial record pushed to a
-// server overwrites the complete copy there.
-func (s *Service) HydrateBestEffort(ctx context.Context, e *Event) error {
-	return s.hydrate(ctx, e, false)
-}
-
-// HydrateSkipUnreadable populates every relation it can and returns the names
-// of the relations it could not load. It never fails.
-//
-// This serves the CLI's ical export --skip-unreadable path. The user chose to
-// keep a backup that misses unreadable relations, so the export names what
-// each record lost. Any path that writes iCal for a push must use Hydrate.
-// A partial record pushed to a server overwrites the complete copy there.
-func (s *Service) HydrateSkipUnreadable(ctx context.Context, e *Event) []string {
-	return s.collect(ctx, e, false).Failed()
-}
-
-// collect loads every relation onto e through one Collector. Hydrate,
-// HydrateBestEffort, and HydrateSkipUnreadable share it, so the relation set
-// stays a single definition.
-func (s *Service) collect(ctx context.Context, e *Event, failFast bool) *hydrate.Collector {
-	c := &hydrate.Collector{Kind: "event", ID: e.ID, FailFast: failFast}
-	hydrate.Rel(ctx, c, &e.Alarms, "alarms", s.ListAlarms)
-	hydrate.Rel(ctx, c, &e.Attendees, "attendees", s.ListAttendees)
-	hydrate.Rel(ctx, c, &e.Attachments, "attachments", s.ListAttachments)
-	hydrate.Rel(ctx, c, &e.Comments, "comments", s.ListComments)
-	hydrate.Rel(ctx, c, &e.Contacts, "contacts", s.ListContacts)
-	hydrate.Rel(ctx, c, &e.Resources, "resources", s.ListResources)
-	hydrate.Rel(ctx, c, &e.Relations, "relations", s.ListRelations)
-	hydrate.Rel(ctx, c, &e.XProperties, "x-properties", s.ListXProperties)
-	return c
-}
-
-func (s *Service) hydrate(ctx context.Context, e *Event, failFast bool) error {
-	return s.collect(ctx, e, failFast).Err()
 }
